test(e2e): split iconShape matrix into per-form spec files
iconShape.spec.ts was the heaviest e2e spec by far (~252s, 339 tests) — the
bottleneck for shard balance (cypress-split balances whole files) and the spec
that overran the V8-coverage `coverageAfter` hook.

Extract the matrix (looks × directions × labelPos, 7 cases each) into a shared
helper and split it across one spec file per form:
- iconShape.spec.ts          → base form (no `form`) + the standalone tests
- iconShape-square.spec.ts   → form: square
- iconShape-circle.spec.ts   → form: circle
- iconShape-rounded.spec.ts  → form: rounded

Each is ~84 tests / ~63s. Behavior-preserving: the describe/it titles and the
generated flowchart code are byte-identical (verified by diff), so every
rendered diagram is unchanged — 339 tests total, same as before.

`clazz` added to the cspell dictionary: `customClazz` was previously only in
*.spec.{ts,js} files (cspell-relaxed via eslint override); it now also appears
in the helper, which is fully spell-checked.

Note: because Argos keys screenshots by their spec-file path, the square/circle/
rounded screenshots move to new paths — Argos will show them as removed+added
on the baseline update though the images are identical.
