+++ cypress/integration/rendering/iconShape-rounded.spec.ts
@@ -1,0 +1,5 @@
+import { registerIconShapeTests } from '../../helpers/iconShapeTests';
+
+// `rounded` form slice of the iconShape matrix. See iconShapeTests.ts and
+// iconShape.spec.ts for why this matrix is split across files.
+registerIconShapeTests(['rounded']);
