+++ cypress/integration/rendering/iconShape-square.spec.ts
@@ -1,0 +1,5 @@
+import { registerIconShapeTests } from '../../helpers/iconShapeTests';
+
+// `square` form slice of the iconShape matrix. See iconShapeTests.ts and
+// iconShape.spec.ts for why this matrix is split across files.
+registerIconShapeTests(['square']);
